Refactor: recoil 폴더 정책 적용
