+++ src/Test.tsx
@@ -1,8 +1,8 @@
 import { useSetRecoilState } from 'recoil';
-import { postModalState } from './atoms/postModalState';
 import ReviewPostModal from './components/ReviewPostModal';
 import BasicPostModal from './components/BasicPostModal';
 import MogakPostModal from './components/MogakPostModal';
+import { postModalState } from './recoil/postModal/atoms';
 
 const Test = () => {
   const setPostModalOpen = useSetRecoilState(postModalState);
